Move get_response_schema_section back to schema_tester
